+++ packages/api-fetch/src/middlewares/preloading.js
@@ -1,4 +1,9 @@
 /**
+ * WordPress dependencies
+ */
+import { normalizePath } from '@wordpress/url';
+
+/**
  * Given a path, returns a normalized path where equal query parameter values
  * will be treated as identical, regardless of order they appear in the original
  * text.
@@ -8,29 +13,7 @@
  * @return {string} Normalized path.
  */
 export function getStablePath( path ) {
-	const splitted = path.split( '?' );
-	const query = splitted[ 1 ];
-	const base = splitted[ 0 ];
-	if ( ! query ) {
-		return base;
-	}
-
-	// 'b=1&c=2&a=5'
-	return (
-		base +
-		'?' +
-		query
-			// [ 'b=1', 'c=2', 'a=5' ]
-			.split( '&' )
-			// [ [ 'b, '1' ], [ 'c', '2' ], [ 'a', '5' ] ]
-			.map( ( entry ) => entry.split( '=' ) )
-			// [ [ 'a', '5' ], [ 'b, '1' ], [ 'c', '2' ] ]
-			.sort( ( a, b ) => a[ 0 ].localeCompare( b[ 0 ] ) )
-			// [ 'a=5', 'b=1', 'c=2' ]
-			.map( ( pair ) => pair.join( '=' ) )
-			// 'a=5&b=1&c=2'
-			.join( '&' )
-	);
+	return normalizePath( path );
 }
 
 /**
