+++ packages/url/src/index.js
@@ -23,3 +23,4 @@ export { safeDecodeURIComponent } from './safe-decode-uri-component';
 export { filterURLForDisplay } from './filter-url-for-display';
 export { cleanForSlug } from './clean-for-slug';
 export { getFilename } from './get-filename';
+export { normalizePath } from './normalize-path';
